+++ CHANGELOG.md
@@ -1,8 +1,18 @@
-# HEAD  (2017-07-26)
+# HEAD  (2017-07-28)
 
 Features:
 
 - Publish lifecycle events to Routemaster (#19)
+
+# v1.8.1 (2017-07-27)
+
+Features:
+
+- Google OAuth supported in Rails 3 and 4 (#54)
+
+Bug fixes:
+
+- Allow client apps to load middleware (#54)
 
 # v1.8.0 (2017-07-26)